Find column by name
Since the column can change position from sheet to sheet, find it by name, instead
of making it fix in the code.

In the impossible eventuality that a translator for another language what to use
this script, it will be quite easy to "de-hardcode" the language column name and
make it a script argument. Right now it's fixed to 'it'.